More packaging cleanups
Cleanup more packaging items to make it MUCH easier to push releases in the future. Also remove all extraneous libs. We only need one set of library dependencies, and those can be in the main addon. All of the modules are now down to 'bare bones', which is much cleaner.