fix rest of warnings
Fixed the rest of the code warnings in the main addon. I shouldn't have functionally changed anything, it should all just be making the code easier to read and easier for the debugger to sort through. Most of the errors have to do with empty or unused variable assignments, or unbalanced variable expressions