Major Performance fixes
Using the AddonUsage addon, I noticed that Neuron was using super high CPU usage, and I was able to track it down to a seeminly unnecessary OnUpdate call in the Neuron_Buttons.lua file. It was calling this function each and every frame update, which shot the CPU usage through the roof. Removing it doesn't seem to harm anything, but I'm skeptical.