added a hack for feral druid thrash and clearcasting
Blizzard has a big bug that doesn't register Thrash as procced on the real spellID, being 106832, but rather shows the proc on some weird joing feral/guardian spellID 106830 instead. The dumb thing is that 10680 is kind of a useless spell, as there is no ability called Thrash(Feral, Guardian) from what I can tell, so why the proc is happening on that spellID is beyond me.