new features and fixes
Add a new feature to the status bars to allow you to see the current rank/level of whatever it is you are tracking, i.e. Player Level, Honor Level, AP Points Spent, etc.
Fixed the Checkbutton not working on the XP Bar
Made it so that you can now have multiple XP bars and their settings won't conflict
Fixed the XP bars not remember their settings upon logging in.