get rid of StanceBars module
This might be a controversial decision, but it was only applicable to druids anyway since Legion, and it was causing problems elsewhere. We might as well just add some logic to the first time startup to make a normal actionbar with the druid forms on it if people deem it a must-have. For now it's gone though