Fix errors caused by items that are not cached
Items that are used in a button's macro via their slot id (i.e. "/use 15" to use cloak) are never cached in the ItemCache table and will therefore cause a concatenation error every few frames. This edit simply performs a nil check and gets icon info if it is not cached.