Refine taint protection, kinda messy very sledgehammery
The way I'm protecting against any and all taint issues relating to ActionButtons is by intercepting the OnEvent call, this is pretty hardcore and not the most elegant solution. The downside to this is that all Extra. ZoneAbility, and Exit buttons are disabled all the time now, not just when there's a Neuron equivalent present