fixes #249 and #266, Autowrite function will check for alt name
There is an issue in game where abilities that have a base spell and then a talented form, i.e Roll and Chi Torpedo, if you drag the talented form directly it won't properly adapt to the untalented version.

This patch makes it so if there is an alt name (i,e. the base spell) for the spell, and the alt spell name is known (which is only possible if it IS the base spell), it will write the macro using that name instead