Initial 9.0.1 port. Lots of changes
had to port the addon, namely the aweful GUI, to the new BackdropTemplate system introducted in 9.0. This took FOREVER to find all the frames needing fixing.
Rewrote initial logic for the ZoneAbility button. Blizzard is expanding this functionality it seems to allow for multiple zone ability buttons at a given time. We will have to deal with that in the future, however this works for now
Lots of our RawHooks for disabling the default functionality now seem unnecessary. They removed 4/7 of the functions we were overriding anyhow. We will need to keep an eye on it to make sure we don't get taint